+++ tests/testthat/test-function-readLinesWithEncoding.R
@@ -1,0 +1,8 @@
+test_that("readLinesWithEncoding() works", {
+
+  expect_error(
+    readLinesWithEncoding()
+    # argument "file" is missing, with no default
+  )
+
+})